// src/app/models/Proveedor.ts
export class Proveedor{
    IdProveedor: number=0;
    RazonSocial: string= "";
    Nombre: string= "";
    IdTributaria: number= 0;
    Telefono: number= 0;
    Correo: string= "";
    SitioWeb: string= "";
    Direccion: string= "";
    Pais: string= "";
    Facturacion: number=0; 
    UltimaEdicion:Date = new Date();
}